Add in dynamic detection of battery in lieu of assuming BAT0
Originally, this project assumed BAT0 was always going to be reported
as the primary battery. A samsung laptop flew in the face of that
assumption, identifying its battery as BAT1, and I will need to
refactor my code to allow dynamic detection of the battery sysfs node.

I will need later on to figure out how to meaningfully accomodate
multi-battery laptops as well, as they are still commonly present.
I also still assume that the lowest-numbered BATx will be the primary
battery that conveys useful power-draw information.

This commit adds an init for battery info that globs the sysfs
power_supply directory for BAT*, and assigns the full path to
to global allocated string, BatteryPath.