Update all battery functions to make use of global var BatteryPath
I am making use of asprintf for source clarity, and I am
intentionally not handling errors on asprintf  because a
condition where malloc fails on asprintf is a failure condition
for the whole application, and I'm happy to just let it segfault.

Now the application should be able to handle circumstances where
BAT0 is not named, and the primary battery actually starts as BAT1
or higher number.