add -Wno-unused-result to CFLAGS to silence asprintf warnings
I am consciously deciding not to handle errors with asprintf. if
malloc fails for operations this tiny, there are problems with
the runtime that cannot be handled gracefully by this application.

From a security perspective, /sys would need to be compromised in
order to manipulate anything asprintf touches right now, and that
is outside the scope of my program to consider.