rewrite add_node to use a cursor pointer
I was being clever when using an argument variable as a throwaway
cursor, but this could give the impression that I am modifying
the input pointer as an intentional output of the function. by using
a separate cursor variable, I consume another 8 bytes of data on the stack
in return for greater clarity of code.