Add a new window that lists every cpu's freqency state
This is a huge commit, but it adds a feature that allows a user to press 'f' on
the keyboard and display a listing of every cpu's current frequency state. The
current window space supports barely more than 128 CPUs, which already limits
its use. I've written the function to parameterize against the window size
should I decide to make it bigger or resize to the terminal.

In its current state, my program is far more useful on intel laptops than any
other computing platform, and I really don't believe laptop cpus will exceed 128
threads for at least 5 or so years, but its worth considering flexibility with
this program, as some crazy bastards just put a 32-thread Ryzen 9 3950X into a
Clevo laptop this year.

Listing of changes made:
-add gethc() and set up ncurses to read cbreak() and clear input buffer upon
reading.
-add a get_cur_freq() function that fetches individual cpu frequencies
-base get_freq() on get_cur_freq(0)
-revise the main application loop to switch modes based upon 'f' key
-create draw_grid() function, which lists all cpu freqencies as a window
-create a test_grid function, where I stress test the program logic that was
implemented in draw_grid()