Add preliminary support for intel Hardware-Controlled Performance
On top of P-state controls, speedstep, Hardware Duty Cycling, and
RAPL power limiting, intel also supports something they call
Hardware-Controlled Performance, or HWP, the most intriguing feature
being the ability for an OS to send an arbitrary number between 0-255
to correlate to how much or little power-savings you want. this is
enabled by default on linux systems with the p-state cpufreq driver
it seems, and modern cpufreq widgets mess with it, so I will too.